remove test utils and integration test for enumeration and short call
