+++ ddk/src/ddk.rs
@@ -242,164 +242,3 @@ where
         Ok((contract_id, counter_party, accept_dlc))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::time::{Duration, SystemTime, UNIX_EPOCH};
-
-    use crate::{
-        test_util::{generate_blocks, test_ddk, TestSuite},
-        Transport,
-    };
-    use ddk_manager::{
-        contract::{contract_input::ContractInput, Contract},
-        Storage,
-    };
-    use ddk_messages::{oracle_msgs::OracleAnnouncement, Message};
-    use rstest::rstest;
-    use tokio::time::sleep;
-
-    #[rstest]
-    #[test_log::test(tokio::test)]
-    async fn contract_execution(
-        #[future] test_ddk: (
-            TestSuite,
-            TestSuite,
-            (u32, OracleAnnouncement),
-            ContractInput,
-        ),
-    ) {
-        let (alice, bob, announcement, contract_input) = test_ddk.await;
-        let (id, announcement) = announcement;
-
-        let alice_makes_offer = alice.ddk.manager.send_offer_with_announcements(
-            &contract_input,
-            bob.ddk.transport.keypair.public_key(),
-            vec![vec![announcement.clone()]],
-        );
-
-        let alice_makes_offer = alice_makes_offer
-            .await
-            .expect("alice did not create an offer");
-
-        let contract_id = alice_makes_offer.temporary_contract_id.clone();
-        let alice_pubkey = alice.ddk.transport.public_key();
-        let bob_pubkey = bob.ddk.transport.public_key();
-
-        let bob_receives_offer = bob
-            .ddk
-            .manager
-            .on_dlc_message(&Message::Offer(alice_makes_offer), alice_pubkey)
-            .await;
-
-        let bob_receive_offer = bob_receives_offer.expect("bob did not receive the offer");
-        assert!(bob_receive_offer.is_none());
-
-        let bob_accept_offer = bob
-            .ddk
-            .manager
-            .accept_contract_offer(&contract_id)
-            .await
-            .expect("bob could not accept offer");
-
-        let (contract_id, _counter_party, bob_accept_dlc) = bob_accept_offer;
-
-        let alice_receive_accept = alice
-            .ddk
-            .manager
-            .on_dlc_message(&Message::Accept(bob_accept_dlc), bob_pubkey)
-            .await
-            .expect("alice did not receive accept");
-
-        assert!(alice_receive_accept.is_some());
-
-        let alice_sign_message = alice_receive_accept.unwrap();
-        bob.ddk
-            .manager
-            .on_dlc_message(&alice_sign_message, alice_pubkey)
-            .await
-            .expect("bob did not receive sign message");
-
-        generate_blocks(10);
-
-        alice
-            .ddk
-            .manager
-            .periodic_check(false)
-            .await
-            .expect("alice check failed");
-
-        bob.ddk
-            .manager
-            .periodic_check(false)
-            .await
-            .expect("bob check failed");
-
-        let contract = alice.ddk.storage.get_contract(&contract_id);
-        assert!(matches!(contract.unwrap().unwrap(), Contract::Confirmed(_)));
-
-        bob.ddk.wallet.sync().unwrap();
-        alice.ddk.wallet.sync().unwrap();
-
-        // Used to check that timelock is reached.
-        let locktime = match alice.ddk.storage.get_contract(&contract_id).unwrap() {
-            Some(contract) => match contract {
-                Contract::Confirmed(signed_contract) => {
-                    signed_contract.accepted_contract.dlc_transactions.cets[0]
-                        .lock_time
-                        .to_consensus_u32()
-                }
-                _ => unreachable!("No locktime."),
-            },
-            None => unreachable!("No locktime"),
-        };
-
-        let mut time = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_secs() as u32;
-
-        let attestation = alice
-            .ddk
-            .oracle
-            .oracle
-            .sign_enum_event(id, "rust".to_string())
-            .await;
-
-        while time < announcement.oracle_event.event_maturity_epoch || time < locktime {
-            tracing::warn!("Waiting for time to expire for oracle event and locktime.");
-            let checked_time = SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .unwrap()
-                .as_secs() as u32;
-
-            time = checked_time;
-            generate_blocks(5);
-        }
-
-        assert!(attestation.is_ok());
-
-        bob.ddk.wallet.sync().unwrap();
-        alice.ddk.wallet.sync().unwrap();
-
-        bob.ddk
-            .manager
-            .close_confirmed_contract(&contract_id, vec![(0, attestation.unwrap())])
-            .await
-            .unwrap();
-
-        sleep(Duration::from_secs(10)).await;
-
-        bob.ddk.manager.periodic_check(false).await.unwrap();
-
-        let contract = bob.ddk.storage.get_contract(&contract_id).unwrap().unwrap();
-        assert!(matches!(contract, Contract::PreClosed(_)));
-
-        generate_blocks(10);
-
-        bob.ddk.manager.periodic_check(false).await.unwrap();
-
-        let contract = bob.ddk.storage.get_contract(&contract_id);
-        assert!(matches!(contract.unwrap().unwrap(), Contract::Closed(_)));
-    }
-}
